pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["src/", "experiments/"]

[project]
name = "visual-question-decomposition"
version = "1.0.0"
description = "Codebase for Analyzing Modular Approaches for Visual Question Decomposition (EMNLP 2023)"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "apoorvkh", email = "mail@apoorvkh.com"},]
dependencies = [
    "ai2-tango==1.3.2",
    "datasets==2.14.4",
    "openai==0.27.7",
    "Pillow==9.4.0",
    "rich==13.5.2",
    "sqlitedict==2.1.0",
    "accelerate==0.18.0",
    "bitsandbytes==0.38.1",
    "transformers==4.29.2",
    "torch==2.1.0",
    "torchvision==0.16.0",
    "tyro==0.6.2",
    "viper @ git+https://github.com/brown-palm/visual-question-decomposition.git@viper"
]

[tool.pdm.dev-dependencies]
dev = [
    "ruff==0.1.5",
]

[tool.ruff]
select = ["E", "F", "I"]
line-length = 100
src = ["src"]
target-version = "py39"

[tool.ruff.isort]
known-first-party = ["src", "experiments"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
